fix(pill): update stories for pill component
