--- core/components/atoms/pills/__stories__/Default.story.jsx
@@ -1,20 +1,0 @@
-import * as React from 'react';
-import Pills from '../../pills';
-
-export const defaultPill = () => (
-  <Pills appearance="secondary" subtle={false}>
-    10
-  </Pills>
-);
-
-export default {
-  title: 'Components/Pills/Default Pill',
-  component: Pills,
-  parameters: {
-    docs: {
-      docPage: {
-        title: 'Pills',
-      },
-    },
-  },
-};
