+++ core/components/atoms/pills/__stories__/Solid.story.jsx
@@ -1,10 +1,9 @@
 import * as React from 'react';
-import Pills from '../../Pills';
-import Text from '@/components/atoms/text';
+import { Pills, Text } from '@/index';
 
 // CSF format story
-export const subtle = () => {
-  const ButtonSubtle = true;
+export const solid = () => {
+  const subtle = false;
   const weight = 'strong';
   const children = 10;
   const appearances = [
@@ -18,12 +17,13 @@ export const subtle = () => {
     'accent3',
     'accent4',
   ];
+
   return (
     <div className="d-flex">
       {appearances.map((appear, ind) => {
         return (
           <div key={ind} className="mr-9">
-            <Pills appearance={appear} subtle={ButtonSubtle}>
+            <Pills appearance={appear} subtle={subtle}>
               {children}
             </Pills>
             <br />
@@ -36,7 +36,7 @@ export const subtle = () => {
 };
 
 export default {
-  title: 'Components/Pills/Variants/Subtle',
+  title: 'Components/Pills/Solid',
   component: Pills,
   parameters: {
     docs: {
