--- core/components/atoms/pills/__stories__/SubtleWarning.story.jsx
@@ -1,20 +1,0 @@
-import * as React from 'react';
-import Pills from '../../pills';
-
-export const subtleWarning = () => (
-  <Pills appearance="warning" subtle={true}>
-    10
-  </Pills>
-);
-
-export default {
-  title: 'Components/Pills/Subtle Warning',
-  component: Pills,
-  parameters: {
-    docs: {
-      docPage: {
-        title: 'Pills',
-      },
-    },
-  },
-};
