+++ core/components/atoms/pills/__stories__/index.story.jsx
@@ -1,17 +1,9 @@
 import * as React from 'react';
-import Pills from '../Pills';
+import { Pills } from '@/index';
 
 // CSF format story
 export const all = () => {
-  const appearance = 'warning';
-  const subtle = false;
-  const children = 10;
-
-  return (
-    <Pills appearance={appearance} subtle={subtle}>
-      {children}
-    </Pills>
-  );
+  return <Pills>10</Pills>;
 };
 
 export default {
